--- meta/main.yml
@@ -1,21 +1,0 @@
----
-galaxy_info:
-  author: "Ogonna Iwunze"
-  company: None
-  license: BSD
-  min_ansible_version: 2.0
-  platforms:
-    - name: EL
-      versions:
-        - 7
-    - name: Ubuntu
-      versions:
-        - precise
-        - trusty
-        - xenial
-  galaxy_tags:
-    - elk
-    - filebeat
-    - logging
-dependencies: []
-
